Move generic utility functions to 'proxysql_utils.h' #3847
Use 'proxysql_utils.h' as an isolated header in which place common
utility functions that can be used from ProxySQL and from outside tests.
